Simplify Core AI provider surface
