+++ packages/core-ai/src/ai-sdk.ts
@@ -7,29 +7,28 @@ import type {
   LanguageModelV3,
   LanguageModelV3CallOptions,
   LanguageModelV3FinishReason,
-  LanguageModelV3Prompt,
   LanguageModelV3StreamPart,
   TranscriptionModelV3,
   TranscriptionModelV3CallOptions,
 } from '@ai-sdk/provider'
 import { generateId } from '@ai-sdk/provider-utils'
+import type { UnsafeObject } from 'react-native/Libraries/Types/CodegenTypes'
 
-import {
-  coreAI,
-  CoreAIEmbeddingModel,
-  CoreAIImageModel,
-  CoreAILanguageModel,
-  CoreAITranscriptionModel,
-  prepareMessages,
-} from './core'
 import NativeCoreAI from './NativeCoreAI'
 import type {
+  CoreAIEmbeddingResult,
   CoreAIGenerationOptions,
   CoreAIGenerationPart,
+  CoreAIImageGenerationOptions,
+  CoreAIImageGenerationResult,
+  CoreAILoadedModel,
+  CoreAIMessage,
   CoreAIModelConfig,
+  CoreAIModelInfo,
   CoreAIStreamCompleteEvent,
   CoreAIStreamErrorEvent,
   CoreAIStreamUpdateEvent,
+  CoreAITranscriptionResult,
 } from './types'
 import { toNativeModelConfig } from './types'
 
@@ -38,9 +37,9 @@ export function createCoreAIProvider() {
     return provider.languageModel(config)
   }
 
-  provider.getCapabilities = () => coreAI.getCapabilities()
+  provider.getCapabilities = () => NativeCoreAI.getCapabilities()
   provider.languageModel = (config: CoreAIModelConfig) => {
-    return new CoreAILanguageModelAdapter(config)
+    return new CoreAILanguageModel(config)
   }
   provider.textEmbeddingModel = (config: CoreAIModelConfig) => {
     return new CoreAITextEmbeddingModel(config)
@@ -50,32 +49,95 @@ export function createCoreAIProvider() {
     return new CoreAIImageGenerationModel(config)
   }
   provider.transcriptionModel = (config: CoreAIModelConfig) => {
-    return new CoreAITranscriptionAdapter(config)
+    return new CoreAITranscriptionModel(config)
   }
 
   return provider
 }
 
-export const coreAIProvider = createCoreAIProvider()
+export const coreAI = createCoreAIProvider()
 
-export class CoreAILanguageModelAdapter
-  extends CoreAILanguageModel
-  implements LanguageModelV3
-{
+export interface CoreAILanguageSession {
+  sessionHandle: string
+  respond(
+    prompt: string,
+    options?: CoreAIGenerationOptions
+  ): Promise<CoreAIGenerationPart[]>
+  stream(
+    prompt: string,
+    options?: CoreAIGenerationOptions
+  ): Promise<ReadableStream<CoreAIStreamUpdateEvent>>
+  close(): Promise<void>
+}
+
+export class CoreAILanguageModel implements LanguageModelV3 {
   readonly specificationVersion = 'v3'
   readonly supportedUrls = {}
   readonly provider = 'core-ai'
   readonly modelId: string
+  readonly config: CoreAIModelConfig
+
+  private loadedModel?: CoreAILoadedModel
 
   constructor(config: CoreAIModelConfig) {
-    super({ ...config, task: 'language' })
+    this.config = { ...config, task: 'language' }
     this.modelId = config.id
+  }
+
+  get modelHandle() {
+    return this.loadedModel?.modelHandle
+  }
+
+  async inspect(): Promise<CoreAIModelInfo> {
+    return NativeCoreAI.inspectModel(
+      toNativeModelConfig(this.config)
+    ) as Promise<CoreAIModelInfo>
+  }
+
+  async prepare(options: UnsafeObject = {}): Promise<CoreAIModelInfo> {
+    const loadedModel = (await NativeCoreAI.loadModel(
+      toNativeModelConfig(this.config),
+      options
+    )) as CoreAILoadedModel
+    this.loadedModel = loadedModel
+    return loadedModel.info
+  }
+
+  async specialize(options: UnsafeObject = {}): Promise<CoreAIModelInfo> {
+    return NativeCoreAI.specializeModel(
+      toNativeModelConfig(this.config),
+      options
+    ) as Promise<CoreAIModelInfo>
+  }
+
+  async unload(): Promise<void> {
+    if (!this.loadedModel) {
+      return
+    }
+    await NativeCoreAI.unloadModel(this.loadedModel.modelHandle)
+    this.loadedModel = undefined
+  }
+
+  async remove(): Promise<void> {
+    await NativeCoreAI.removeModel(toNativeModelConfig(this.config))
+    this.loadedModel = undefined
+  }
+
+  async createSession(
+    options: UnsafeObject = {}
+  ): Promise<CoreAILanguageSession> {
+    const model = await this.ensureLoaded()
+    const sessionHandle = await NativeCoreAI.createLanguageSession(
+      model.modelHandle,
+      options
+    )
+    return createLanguageSession(sessionHandle)
   }
 
   async doGenerate(options: LanguageModelV3CallOptions) {
     const response = (await NativeCoreAI.generateText(
       toNativeModelConfig(this.config),
-      toCoreAIMessages(options.prompt),
+      prepareMessages(options.prompt),
       toCoreAIGenerationOptions(options)
     )) as CoreAIGenerationPart[]
 
@@ -148,7 +210,7 @@ export class CoreAILanguageModelAdapter
         NativeCoreAI.streamText(
           streamId,
           toNativeModelConfig(this.config),
-          toCoreAIMessages(options.prompt),
+          prepareMessages(options.prompt),
           toCoreAIGenerationOptions(options)
         )
       },
@@ -166,21 +228,37 @@ export class CoreAILanguageModelAdapter
       },
     }
   }
+
+  private async ensureLoaded(): Promise<CoreAILoadedModel> {
+    if (!this.loadedModel) {
+      await this.prepare()
+    }
+    return this.loadedModel!
+  }
 }
 
-export class CoreAITextEmbeddingModel
-  extends CoreAIEmbeddingModel
-  implements EmbeddingModelV3
-{
+export class CoreAITextEmbeddingModel implements EmbeddingModelV3 {
   readonly specificationVersion = 'v3'
   readonly provider = 'core-ai'
   readonly modelId: string
   readonly maxEmbeddingsPerCall = Infinity
   readonly supportsParallelCalls = false
+  readonly config: CoreAIModelConfig
 
   constructor(config: CoreAIModelConfig) {
-    super({ ...config, task: 'embedding' })
+    this.config = { ...config, task: 'embedding' }
     this.modelId = config.id
+  }
+
+  async embed(
+    values: string[],
+    options: UnsafeObject = {}
+  ): Promise<CoreAIEmbeddingResult> {
+    return NativeCoreAI.embed(
+      toNativeModelConfig(this.config),
+      values,
+      options
+    ) as Promise<CoreAIEmbeddingResult>
   }
 
   async doEmbed(
@@ -194,18 +272,27 @@ export class CoreAITextEmbeddingModel
   }
 }
 
-export class CoreAIImageGenerationModel
-  extends CoreAIImageModel
-  implements ImageModelV3
-{
+export class CoreAIImageGenerationModel implements ImageModelV3 {
   readonly specificationVersion = 'v3'
   readonly provider = 'core-ai'
   readonly modelId: string
   readonly maxImagesPerCall = 1
+  readonly config: CoreAIModelConfig
 
   constructor(config: CoreAIModelConfig) {
-    super({ ...config, task: 'diffusion' })
+    this.config = { ...config, task: 'diffusion' }
     this.modelId = config.id
+  }
+
+  async generate(
+    prompt: string,
+    options: CoreAIImageGenerationOptions = {}
+  ): Promise<CoreAIImageGenerationResult> {
+    return NativeCoreAI.generateImage(
+      toNativeModelConfig(this.config),
+      prompt,
+      options
+    ) as Promise<CoreAIImageGenerationResult>
   }
 
   async doGenerate(options: ImageModelV3CallOptions) {
@@ -237,17 +324,28 @@ export class CoreAIImageGenerationModel
   }
 }
 
-export class CoreAITranscriptionAdapter
-  extends CoreAITranscriptionModel
-  implements TranscriptionModelV3
-{
+export class CoreAITranscriptionModel implements TranscriptionModelV3 {
   readonly specificationVersion = 'v3'
   readonly provider = 'core-ai'
   readonly modelId: string
+  readonly config: CoreAIModelConfig
 
   constructor(config: CoreAIModelConfig) {
-    super({ ...config, task: 'asr' })
+    this.config = { ...config, task: 'asr' }
     this.modelId = config.id
+  }
+
+  async transcribe(
+    audio: Uint8Array | string,
+    mediaType: string,
+    options: UnsafeObject = {}
+  ): Promise<CoreAITranscriptionResult> {
+    return NativeCoreAI.transcribe(
+      toNativeModelConfig(this.config),
+      toBase64(audio),
+      mediaType,
+      options
+    ) as Promise<CoreAITranscriptionResult>
   }
 
   async doGenerate(options: TranscriptionModelV3CallOptions) {
@@ -274,8 +372,108 @@ export class CoreAITranscriptionAdapter
   }
 }
 
-function toCoreAIMessages(prompt: LanguageModelV3Prompt) {
-  return prepareMessages(prompt)
+function createLanguageSession(sessionHandle: string): CoreAILanguageSession {
+  return {
+    sessionHandle,
+    respond(prompt, options = {}) {
+      return NativeCoreAI.respondToLanguageSession(
+        sessionHandle,
+        prompt,
+        options
+      ) as Promise<CoreAIGenerationPart[]>
+    },
+    stream(prompt, options = {}) {
+      return streamLanguageSession(sessionHandle, prompt, options)
+    },
+    close() {
+      return NativeCoreAI.releaseLanguageSession(sessionHandle)
+    },
+  }
+}
+
+function streamLanguageSession(
+  sessionHandle: string,
+  prompt: string,
+  options: CoreAIGenerationOptions
+): Promise<ReadableStream<CoreAIStreamUpdateEvent>> {
+  if (typeof ReadableStream === 'undefined') {
+    throw new Error(
+      'ReadableStream is not available. Load a web stream polyfill before streaming Core AI responses.'
+    )
+  }
+
+  const streamId = generateId()
+  let listeners: { remove(): void }[] = []
+
+  const cleanup = () => {
+    listeners.forEach((listener) => listener.remove())
+    listeners = []
+  }
+
+  const stream = new ReadableStream<CoreAIStreamUpdateEvent>({
+    start(controller) {
+      const updateListener = NativeCoreAI.onStreamUpdate(
+        (event: CoreAIStreamUpdateEvent) => {
+          if (event.streamId === streamId) {
+            controller.enqueue(event)
+          }
+        }
+      )
+      const completeListener = NativeCoreAI.onStreamComplete(
+        (event: CoreAIStreamCompleteEvent) => {
+          if (event.streamId === streamId) {
+            cleanup()
+            controller.close()
+          }
+        }
+      )
+      const errorListener = NativeCoreAI.onStreamError(
+        (event: CoreAIStreamErrorEvent) => {
+          if (event.streamId === streamId) {
+            cleanup()
+            controller.error(new Error(event.error))
+          }
+        }
+      )
+
+      listeners = [updateListener, completeListener, errorListener]
+      NativeCoreAI.streamLanguageSession(
+        streamId,
+        sessionHandle,
+        prompt,
+        options
+      )
+    },
+    cancel() {
+      cleanup()
+      NativeCoreAI.cancelStream(streamId)
+    },
+  })
+
+  return Promise.resolve(stream)
+}
+
+export function prepareMessages(messages: unknown): CoreAIMessage[] {
+  if (!Array.isArray(messages)) {
+    return []
+  }
+
+  return messages.map((message: any): CoreAIMessage => {
+    const content = Array.isArray(message.content)
+      ? message.content.reduce((acc: string, part: any) => {
+          if (part.type === 'text') {
+            return acc + part.text
+          }
+          console.warn('Unsupported Core AI message content type:', part)
+          return acc
+        }, '')
+      : String(message.content ?? '')
+
+    return {
+      role: message.role,
+      content,
+    }
+  })
 }
 
 function toCoreAIGenerationOptions(
@@ -345,4 +543,18 @@ function emptyUsage() {
       reasoning: undefined,
     },
   }
+}
+
+function toBase64(data: Uint8Array | string): string {
+  if (typeof data === 'string') {
+    return data
+  }
+
+  let binary = ''
+  const chunkSize = 0x8000
+  for (let i = 0; i < data.length; i += chunkSize) {
+    const chunk = data.subarray(i, i + chunkSize)
+    binary += String.fromCharCode(...chunk)
+  }
+  return btoa(binary)
 }
